Split dotgeneral.go into dotgeneral_large.go and dotgeneral_small.go. Return immediately intermediary block-aligned buffers to the pool.
